Annual Leave Entitlement entity added

+++ Core/IdeKusgozManagement.Application/DTOs/UserDTOs/AnnualLeaveCalculationDTO.cs
@@ -2,7 +2,7 @@
 {
     public class AnnualLeaveCalculationDTO
     {
-        public double UsedDays { get; set; }
-        public double RemainingDays { get; set; }
+        public decimal UsedDays { get; set; }
+        public decimal RemainingDays { get; set; }
     }
 }
